# openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
...
    /api/resource:
        get:
            tags:
                - Resource
            operationId: Resource_GetResource
            parameters:
                - name: id
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.yyets.v1.GetResourceReply'
    /api/resource/search:
        get:
            tags:
                - Resource
            operationId: Resource_SearchResource
            parameters:
                - name: keyword
                  in: query
                  schema:
                    type: string
                - name: type
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.yyets.v1.SearchResourceReply'
...
components:
    schemas:
...
        api.yyets.v1.GetResourceReply:
            type: object
            properties:
                id:
                    type: string
                cnname:
                    type: string
                enname:
                    type: string
                aliasname:
                    type: string
                area:
                    type: string
                views:
                    type: string
                data:
                    $ref: '#/components/schemas/api.yyets.v1.ResourceData'
...
        api.yyets.v1.ResourceData:
            type: object
            properties:
                info:
                    $ref: '#/components/schemas/api.yyets.v1.ResourceDataInfo'
                list:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.yyets.v1.ResourceSeason'
        api.yyets.v1.ResourceDataInfo:
            type: object
            properties:
                id:
                    type: string
                cnname:
                    type: string
                enname:
                    type: string
                aliasname:
                    type: string
                channel:
                    type: string
                channelCn:
                    type: string
                area:
                    type: string
                showType:
                    type: string
                expire:
                    type: string
                views:
                    type: string
                year:
                    type: array
                    items:
                        type: string
        api.yyets.v1.ResourceFile:
            type: object
            properties:
                way:
                    type: string
                wayCn:
                    type: string
                address:
                    type: string
                passwd:
                    type: string
...
        api.yyets.v1.ResourceSeason:
            type: object
            properties:
                seasonNum:
                    type: string
                seasonCn:
                    type: string
                items:
                    type: object
                    additionalProperties:
                        $ref: '#/components/schemas/api.yyets.v1.ResourceSeasonList'
                formats:
                    type: array
                    items:
                        type: string
        api.yyets.v1.ResourceSeasonItem:
            type: object
            properties:
                itemId:
                    type: string
                episode:
                    type: string
                name:
                    type: string
                size:
                    type: string
                yyetsTrans:
                    type: string
                dateline:
                    type: string
                files:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.yyets.v1.ResourceFile'
        api.yyets.v1.ResourceSeasonList:
            type: object
            properties:
                items:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.yyets.v1.ResourceSeasonItem'
        api.yyets.v1.SearchComment:
            type: object
            properties:
                username:
                    type: string
                date:
                    type: string
                comment:
                    type: string
                commentID:
                    type: string
                resourceID:
                    type: string
                resourceName:
                    type: string
                origin:
                    type: string
                hasAvatar:
                    type: boolean
                hash:
                    type: string
        api.yyets.v1.SearchResourceReply:
            type: object
            properties:
                resource:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.yyets.v1.ResourceDataInfo'
                comment:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.yyets.v1.SearchComment'
                subtitle:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.yyets.v1.SearchSubtitle'
        api.yyets.v1.SearchSubtitle:
            type: object
            properties: {}
...